[Build] don't cast validation i/o data to float32 while loading

+++ src/finn/builder/build_dataflow_config.py
@@ -389,6 +389,6 @@ class DataflowBuildConfig:
             )
             verify_expected_output_npy = np.load(self.verify_expected_output_npy)
             return (
-                verify_input_npy.astype(np.float32),
-                verify_expected_output_npy.astype(np.float32),
+                verify_input_npy,
+                verify_expected_output_npy,
             )
